// src/generated/models/applicationDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Carmentis indexer API
 * Documentation for the Carmentis indexer API.
 * OpenAPI spec version: 1.0
 */

export interface ApplicationDto {
  virtualBlockchainId: string;
  organizationId: string;
  name: string;
  logoUrl: string;
  homepageUrl: string;
  description: string;
}