// src/generated/models/organizationDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Carmentis indexer API
 * Documentation for the Carmentis indexer API.
 * OpenAPI spec version: 1.0
 */

export interface OrganizationDto {
  virtualBlockchainId: string;
  accountId: string;
  name: string;
  city: string;
  countryCode: string;
  website: string;
}